auth: extract token synthesis and credential validators
Move the two families of pure, stateless helpers out of the God-file
verbatim: sandboxTokenFor (opaque ID-token string) into sandbox-token.ts,
and validateEmailFormat / validatePasswordStrength into
sandbox-credential-validators.ts. Both gain 'export' and the core imports
them; no re-export needed (they were never part of the barrel's public
surface). Bodies verbatim.
